items/views.py
from rest_framework import viewsets

from items import serializers, models


class ItemViewSet(viewsets.ModelViewSet):
    queryset = models.Item.objects.all()
    serializer_class = serializers.ItemSerializer
